Simplify module structure by not re-exporting from `Plutarch.Internal.Other`
