+++ Plutarch.hs
@@ -1,46 +1,49 @@
 module Plutarch (
-  (:-->),
-  ClosedTerm,
-  compile,
-  Dig,
-  hashTerm,
-  papp,
-  pdelay,
-  PDelayed,
-  perror,
-  pforce,
-  phoistAcyclic,
-  plam',
-  plet,
-  Term,
-  S,
-  PType,
-  PlutusType (..),
-  printTerm,
-  printScript,
-  (#$),
-  (#),
-  pinl,
-  PCon,
-  PMatch,
-  pcon,
-  pmatch,
-  pto,
-  pfix,
-  POpaque (..),
-  popaque,
-  plam,
-  DerivePNewtype (DerivePNewtype),
+  (PI.:-->),
+  PI.ClosedTerm,
+  PI.compile,
+  PI.Dig,
+  PI.hashTerm,
+  PI.papp,
+  PI.pdelay,
+  PI.PDelayed,
+  PI.perror,
+  PI.pforce,
+  PI.phoistAcyclic,
+  PI.plet,
+  PI.Term,
+  PI.S,
+  PI.PType,
+  PP.PlutusType,
+  PP.PInner,
+  PP.pcon,
+  PP.pmatch,
+  PP.PCon,
+  PP.PMatch,
+  PO.printTerm,
+  PO.printScript,
+  (PL.#$),
+  (PL.#),
+  PL.pinl,
+  PO.pto,
+  PO.pfix,
+  PO.POpaque (PO.POpaque),
+  PO.popaque,
+  PL.plam,
+  PO.DerivePNewtype (PO.DerivePNewtype),
   PT.TermCont (TermCont),
   PT.hashOpenTerm,
   PT.runTermCont,
   PT.unTermCont,
-  Config (Config, tracingMode),
-  TracingMode (NoTracing, DoTracing, DetTracing),
-  pgetConfig,
-  defaultConfig,
+  PI.Config (Config, tracingMode),
+  PI.TracingMode (NoTracing, DoTracing, DetTracing),
+  PI.pgetConfig,
+  PI.defaultConfig,
 ) where
 
-import Plutarch.Internal.Other
+import qualified Plutarch.Internal as PI
+import qualified Plutarch.Internal.Other as PO
+import qualified Plutarch.Internal.PLam as PL
+import qualified Plutarch.Internal.PlutusType as PP
 import qualified Plutarch.TermCont as PT
 import Prelude ()
